renaming 'mapper' package to 'detectormapper'
