+++ anomdetect/src/main/java/com/expedia/adaptivealerting/anomdetect/comp/DefaultDetectorSource.java
@@ -18,7 +18,7 @@ package com.expedia.adaptivealerting.anomdetect.comp;
 import com.expedia.adaptivealerting.anomdetect.comp.connector.ModelServiceConnector;
 import com.expedia.adaptivealerting.anomdetect.comp.legacy.LegacyDetectorFactory;
 import com.expedia.adaptivealerting.anomdetect.detector.Detector;
-import com.expedia.adaptivealerting.anomdetect.mapper.DetectorMatchResponse;
+import com.expedia.adaptivealerting.anomdetect.detectormapper.DetectorMatchResponse;
 import com.expedia.metrics.MetricDefinition;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
