+++ anomdetect/src/main/java/com/expedia/adaptivealerting/anomdetect/detectormapper/Detector.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.expedia.adaptivealerting.anomdetect.mapper;
+package com.expedia.adaptivealerting.anomdetect.detectormapper;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
