+++ anomdetect/src/main/java/com/expedia/adaptivealerting/anomdetect/detectormapper/DetectorMatchResponse.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.expedia.adaptivealerting.anomdetect.mapper;
+package com.expedia.adaptivealerting.anomdetect.detectormapper;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
