+++ anomdetect/src/test/java/com/expedia/adaptivealerting/anomdetect/comp/connector/ModelServiceConnectorTest.java
@@ -18,8 +18,8 @@ package com.expedia.adaptivealerting.anomdetect.comp.connector;
 import com.expedia.adaptivealerting.anomdetect.DetectorDeserializationException;
 import com.expedia.adaptivealerting.anomdetect.DetectorNotFoundException;
 import com.expedia.adaptivealerting.anomdetect.DetectorRetrievalException;
-import com.expedia.adaptivealerting.anomdetect.mapper.Detector;
-import com.expedia.adaptivealerting.anomdetect.mapper.DetectorMatchResponse;
+import com.expedia.adaptivealerting.anomdetect.detectormapper.Detector;
+import com.expedia.adaptivealerting.anomdetect.detectormapper.DetectorMatchResponse;
 import com.expedia.metrics.MetricDefinition;
 import com.expedia.metrics.TagCollection;
 import com.expedia.metrics.metrictank.MetricTankIdFactory;
