+++ anomdetect/src/test/java/com/expedia/adaptivealerting/anomdetect/detectormapper/CacheUtilTest.java
@@ -1,4 +1,4 @@
-package com.expedia.adaptivealerting.anomdetect.mapper;
+package com.expedia.adaptivealerting.anomdetect.detectormapper;
 
 import org.junit.Assert;
 import org.junit.Test;
