+++ kafka/src/main/java/com/expedia/adaptivealerting/kafka/processor/MetricDataTransformerSupplier.java
@@ -15,8 +15,8 @@
  */
 package com.expedia.adaptivealerting.kafka.processor;
 
-import com.expedia.adaptivealerting.anomdetect.DetectorMapper;
-import com.expedia.adaptivealerting.anomdetect.mapper.MapperResult;
+import com.expedia.adaptivealerting.anomdetect.detectormapper.DetectorMapper;
+import com.expedia.adaptivealerting.anomdetect.detectormapper.MapperResult;
 import com.expedia.metrics.MetricData;
 import lombok.Data;
 import lombok.NonNull;
