+++ kafka/src/test-integration/scala/com/expedia/adaptivealerting/pipeline/integration/test/ConstantThresholdBasedE2ETestSpec.scala
@@ -22,7 +22,8 @@ import com.expedia.adaptivealerting.anomdetect.comp.connector.{HttpClientWrapper
 import com.expedia.adaptivealerting.anomdetect.source.HttpClientWrapper
 import com.expedia.adaptivealerting.anomdetect.source.util.ModelServiceConnector
 import com.expedia.adaptivealerting.anomdetect.util.ModelServiceConnector
-import com.expedia.adaptivealerting.anomdetect.{DetectorManager, DetectorMapper}
+import com.expedia.adaptivealerting.anomdetect.DetectorManager
+import com.expedia.adaptivealerting.anomdetect.detectormapper.DetectorMapper
 import com.expedia.adaptivealerting.core.anomaly.AnomalyResult
 import com.expedia.adaptivealerting.kafka.KafkaConfigProps._
 import com.expedia.adaptivealerting.kafka.KafkaAnomalyDetectorManager
